[TEST] Every function gets its own api test.
