+++ test/api/CMakeLists.txt
@@ -1,4 +1,8 @@
 cmake_minimum_required (VERSION 3.9)
 
 # The api_test build target.
-add_api_test (test_needle.cpp)
+add_api_test (count_test.cpp)
+add_api_test (estimate_test.cpp)
+add_api_test (ibf_test.cpp)
+add_api_test (ibfmin_test.cpp)
+add_api_test (minimiser_test.cpp)
