+++ test/api/count_test.cpp
@@ -1,0 +1,68 @@
+#include <gtest/gtest.h>
+#include <iostream>
+
+#include <seqan3/test/expect_range_eq.hpp>
+
+#include "ibf.h"
+#include "minimiser.h"
+
+#ifndef DATA_INPUT_DIR
+#  define DATA_INPUT_DIR @DATA_INPUT_DIR@
+#endif
+
+using seqan3::operator""_shape;
+std::filesystem::path tmp_dir = std::filesystem::temp_directory_path(); // get the temp directory
+
+void initialization_args(arguments & args)
+{
+    args.compressed = true;
+    args.k = 4;
+    args.shape = seqan3::ungapped{args.k};
+    args.w_size = seqan3::window_size{4};
+    args.s = seqan3::seed{0};
+    args.path_out = tmp_dir/"Test_";
+}
+
+TEST(count, small_example)
+{
+    arguments args{};
+    initialization_args(args);
+
+    count(args, {std::string(DATA_INPUT_DIR) + "mini_example.fasta"}, std::string(DATA_INPUT_DIR) + "mini_gen.fasta",
+          false);
+
+    std::ifstream output_file(tmp_dir/"mini_example.count.out");
+    std::string line;
+    std::string expected{"gen1\t3"};
+    if (output_file.is_open())
+    {
+        while ( std::getline (output_file,line) )
+        {
+            EXPECT_EQ(expected,line);
+        }
+        output_file.close();
+    }
+    std::filesystem::remove(tmp_dir/"mini_example.count.out");
+}
+
+TEST(count, small_example_paired)
+{
+    arguments args{};
+    initialization_args(args);
+
+    count(args, {std::string(DATA_INPUT_DIR) + "mini_example.fasta", std::string(DATA_INPUT_DIR) + "mini_example.fasta"},
+          std::string(DATA_INPUT_DIR) + "mini_gen.fasta", true);
+
+    std::ifstream output_file(tmp_dir/"mini_example.count.out");
+    std::string line;
+    std::string expected{"gen1\t6"};
+    if (output_file.is_open())
+    {
+        while ( std::getline (output_file,line) )
+        {
+            EXPECT_EQ(expected,line);
+        }
+        output_file.close();
+    }
+    std::filesystem::remove(tmp_dir/"mini_example.count.out");
+}
